closes #278: add NULL check to char* logging
Implemented through the macro `ELPP_SIMPLE_LOG_PTR` which is the same as
`ELPP_SIMPLE_LOG`, but does a NULL pointer check.  The auto-space is not
included if the pointer check fails.

void* loggin intentionally does not use `ELPP_SIMPLE_LOG_PTR` as it should
log a raw pointer value.

Added unit test, which fails without the fix (under gcc 4.8.1).
This test in `issues-test.h` as a place for future issues.