Better implementation of string to level/type
The existing implementation has a few problems:

- Duplicating the string in upper and lowercase is really duplicate data. DRY!
- Typing a config string twice in different cases is error prone.
- The user should not be limited to all lowercase or all uppercase. Case shouldn’t matter at all.
- All of the above problems only exist because case-insensitive string comparing is not being used.

This implementation simplifies the whole thing considerably. An existing method, el::base::utils::Str::cStringCaseEq, is used to make a case-insensitive comparison, so the string constant only has to be typed once and case no longer matters. I used lowercase in the constants because it’s easier to read.

The same technique is now applied to converting strings to levels.

Note that I switched to a much more efficient implementation of cStringCaseEq from cygwin. The existing implementation was quite inefficient, at it did a strlen on both strings, which means both strings were traversed completely before they were traversed again for comparison.