Add Emscripten detection, with test modifications
Emscripten allows you to compile C/C++ code to Javascript or WebAssembly
and run it in the browser (or on NodeJS). We can essentially treat it as
a Unix OS, although it needs its own define and a special flag for the
tests to allocate enough memory.

Note that under emscripten some tests don't make sense. You can't shell
out for a start, so the bash helper test is disabled.

The filesystem is also strange under Emscripten - files seem to always
be reported as existing, and there's no restriction on what can be
created. This is reflected in the relevant tests. I suspect anyone using
easylogging++ on the browser wouldn't much care about logging to a file
anyway.

I made these changes in order to integrate with a toy game engine I've
written, and I've tested the emscripten support "in production" with a
game written for Ludum Dare 41, which uses easylogging++ in the browser.

At the time of this commit, the same tests pass that pass when run
natively on my machine. Performance is noticeably less, but that's to be
expected. Runtime performance has, anecdotally, been acceptable.

I absolutely believe that there could be more done to better support
Emscripten with elpp - for now, this will allow elpp to actually compile
and be used in the browser (as the platform detection fails without
these changes)