bug - access attempt to released memory
On the UGW platform, a SIGSEGV event is sent to the APM application
that utilizes the easylogging++ library. The event is sent due to an
access attempt to a released memory. The released memory is a pointer
to the "LogStreamsReferenceMap m_logStreamsReference" which is defined
statically in RegisteredLoggers class.

The RegisteredLoggers object is released in UGW before the release of
other Class objects that make use of the pointer, and thus - the
pointer is accessed after the map is already released.

The issue is not seen in other platforms due to difference in GCC which
results in a difference in object release order.

After this change, the map is defined as a shared-pointer in the
RegisteredLoggers class and is set to a shared-pointer of type
LogStreamsReferenceMap in the RegisteredLoggers CTOR. Even if the
RegisteredLoggers object is released, the map reasouce will be
available to rest of objects until all are destructed.

Signed-off-by: Adam Dov <adamx.dov@intel.com>